+++ src/main/java/com/braintreegateway/DisputeStatusHistory.java
@@ -8,18 +8,24 @@ import com.braintreegateway.util.NodeWrapper;
 
 public final class DisputeStatusHistory {
 
+    private final Calendar disbursementDate;
     private final Calendar effectiveDate;
     private final Calendar timestamp;
     private final Dispute.Status status;
 
     public DisputeStatusHistory(NodeWrapper node) {
-        effectiveDate = node.findDate("timestamp");
+        disbursementDate = node.findDate("disbursement-date");
+        effectiveDate = node.findDate("effective-date");
         timestamp = node.findDateTime("timestamp");
         status = EnumUtils.findByName(Dispute.Status.class, node.findString("status"), Status.UNRECOGNIZED);
     }
 
     public Calendar getEffectiveDate() {
         return effectiveDate;
+    }
+
+    public Calendar getDisbursementDate() {
+        return disbursementDate;
     }
 
     public Calendar getTimestamp() {