+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Add support for `OAUTH_ACCESS_REVOKED` in `WebhookNotification`s
+* Add support for dispute search by `customerId`, `disbursementDate`, and `effectiveDate`
 
 ## 2.77.0
 * Fix possible NullPointerException in UsBankAccount